local-tts-service/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
python-multipart==0.0.24
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
chatterbox-tts==0.1.7
peft>=0.8.0
